Split data access class into partials based
