+++ AllReadyApp/Web-App/AllReady/wwwroot/_references.js
@@ -11,3 +11,11 @@
 /// <reference path="js/site.js" />
 /// <reference path="js/tenant.js" />
 /// <reference path="js/tenants.js" />
+/// <reference path="lib/bootstrap/dist/js/bootstrap.js" />
+/// <reference path="lib/bootstrap-datepicker/js/bootstrap-datepicker.js" />
+/// <reference path="lib/bootstrap-touch-carousel/dist/js/bootstrap-touch-carousel.js" />
+/// <reference path="lib/hammer.js/hammer.js" />
+/// <reference path="lib/jquery/dist/jquery.js" />
+/// <reference path="lib/jquery-validation-unobtrusive/jquery.validate.unobtrusive.js" />
+/// <reference path="lib/knockout/dist/knockout.js" />
+/// <reference path="lib/moment/moment.js" />
